[#182] Refactor: useDropDown 커스텀 훅
